+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Fixed
+* Fixed an issue where map loading would not include some additional models in some cases.
+
 ## 0.21.0 - 2025-11-25
 ### Added
 * Added support for rendering 3D volume textures.